fix(little_test_runner): should not have dependencies, include subpub and copycopy
