+++ clients/webclients/home/development/tests.ts
@@ -1,9 +1,7 @@
 // brian taylor vann
 // home - tests
 
-import { getResults } from "../../utils/little_test_runner/little_test_runner";
+import { runTests } from "../../utils/little_test_runner/little_test_runner";
 
-const results = getResults();
 console.log("tests!");
 console.log("more tests!");
-console.log(results);
