+++ clients/webclients/utils/func/func.ts
@@ -1,0 +1,3 @@
+import { copycopy } from "./copycopy/copycopy";
+
+export { copycopy };
