+++ clients/webclients/utils/little_test_runner/results_store/state_store/actions_types.ts
@@ -1,7 +1,5 @@
-// little test runner
 // brian taylor vann
 
-// Actions
 type StartTestRunActionParams = {
   startTime: number;
 };
@@ -61,6 +59,7 @@ type EndTestRunAction = {
   action: "END_TEST_RUN";
   params: EndTestRunParams;
 };
+
 type ResultsStoreAction =
   | StartTestRunAction
   | StartTestCollectionAction
@@ -70,43 +69,12 @@ type ResultsStoreAction =
   | EndTestCollectionAction
   | EndTestRunAction;
 
-type Consolidate = (action: ResultsStoreAction) => void;
-
-const START_TEST_RUN = "START_TEST_RUN";
-const START_TEST_COLLECTION = "START_TEST_COLLECTION";
-const START_TEST = "START_TEST";
-const CANCEL_RUN = "CANCEL_RUN";
-const END_TEST = "END_TEST";
-const END_TEST_COLLECTION = "END_TEST_COLLECTION";
-const END_TEST_RUN = "END_TEST_RUN";
-
-const consolidate: Consolidate = (action) => {
-  // send to results store
-  switch (action.action) {
-    case START_TEST_RUN:
-      break;
-    case START_TEST_COLLECTION:
-      break;
-    case START_TEST:
-      break;
-    case CANCEL_RUN:
-      break;
-    case END_TEST:
-      break;
-    case END_TEST_COLLECTION:
-      break;
-    case END_TEST_RUN:
-      break;
-    default:
-      break;
-  }
-
-  // dispatch to pubsub
+export {
+  StartTestRunActionParams,
+  StartTestCollectionActionParams,
+  StartTestActionParams,
+  EndTestActionParams,
+  EndTestCollectionActionParams,
+  EndTestRunParams,
+  ResultsStoreAction,
 };
-
-const dispatch: Consolidate = (action) => {
-  consolidate(action);
-  //
-};
-
-export { dispatch };
