+++ clients/webclients/utils/little_test_runner/runner/runner.ts
@@ -1,14 +1,16 @@
 // little test runner
 // brian taylor vann
 
-import { Assertions } from "../results_store/results_store";
-import { Test, Tests, TestCollections } from "../test_store/test_store";
+import {
+  Assertions,
+  Test,
+  TestParams,
+  TestCollection,
+} from "../runner/test_types/test_types";
 import {
   startTestRun,
   startTestCollection,
-  startTest,
   cancelRun,
-  sendTestResult,
   endTestCollection,
   endTestRun,
 } from "./relay_results/relay_results";
@@ -16,149 +18,37 @@ import {
   getTimestamp,
   updateTimestamp,
 } from "./timestamp_sieve/timestamp_sieve";
+import { runTestsInOrder, runTestsAllAtOnce } from "./run_tests/run_tests";
 
-type CreateTestTimeout = (requestedInterval?: number) => Promise<Assertions>;
-type LtrTest = () => Promise<void>;
-type BuildLtrTestParams = {
-  collectionID: number;
-  issuedAt: number;
-  testFunc: Test;
-  testID: number;
-  timeoutInterval?: number;
-};
-type BuildLtrTest = (params: BuildLtrTestParams) => LtrTest;
-interface RunTestsParams {
-  timeoutInterval?: number;
-  tests: Tests;
-  collectionID: number;
+interface StartLtrTestCollectionRunParams {
+  testCollection: TestCollection;
   startTime: number;
 }
-type RunTests = (params: RunTestsParams) => Promise<void>;
-interface RunTestCollectionsParams {
-  testCollections: TestCollections;
+type StartLtrTestCollectionRun = (
+  params: StartLtrTestCollectionRunParams
+) => Promise<void>;
+
+interface StartLtrTestRunParams {
+  testCollection: TestCollection;
 }
-type RunTestCollections = (params: RunTestCollectionsParams) => Promise<void>;
+type StartLtrTestRun = (params: StartLtrTestRunParams) => Promise<void>;
 
 // create a test collection
 
-const sleep = async (time: number) => {
-  return new Promise((resolve) => {
-    setTimeout(() => {
-      resolve();
-    }, time);
-  });
-};
-
-const defaultTimeoutInterval = 10000;
-const getTimeoutAssertions = (timeoutInterval: number) => [
-  `timed out at: ${timeoutInterval}`,
-];
-
-const createTestTimeout: CreateTestTimeout = async (
-  timeoutInterval?: number
-) => {
-  const interval = timeoutInterval ?? defaultTimeoutInterval;
-  await sleep(interval);
-  return getTimeoutAssertions(interval);
-};
-
-const buildTest: BuildLtrTest = (params) => {
-  const { issuedAt, testID, collectionID, timeoutInterval } = params;
-  return async () => {
-    if (issuedAt < getTimestamp()) {
-      return;
-    }
-
-    const startTime = performance.now();
-    startTest({
-      collectionID,
-      testID,
-      startTime,
-    });
-
-    const assertions = await Promise.race([
-      params.testFunc(),
-      createTestTimeout(timeoutInterval),
-    ]);
-
-    if (issuedAt < getTimestamp()) {
-      return;
-    }
-    const endTime = performance.now();
-    sendTestResult({
-      startTime,
-      endTime,
-      assertions,
-      collectionID,
-      testID,
-    });
-  };
-};
-
-const asynchonouslyRunTests: RunTests = async ({
+const startLtrTestCollectionRun: StartLtrTestCollectionRun = async ({
+  testCollection,
   startTime,
-  collectionID,
-  tests,
-  timeoutInterval,
 }) => {
-  const builtAsyncTests = [];
-
-  let testID = 0;
-  for (const testFunc of tests) {
-    builtAsyncTests.push(
-      buildTest({
-        collectionID,
-        issuedAt: startTime,
-        testFunc,
-        testID,
-        timeoutInterval,
-      })
-    );
-    testID += 1;
-  }
-
-  if (startTime < getTimestamp()) {
-    return;
-  }
-  await Promise.all(builtAsyncTests);
-};
-
-const iterateThroughTests: RunTests = async ({
-  startTime,
-  collectionID,
-  tests,
-  timeoutInterval,
-}) => {
-  let testID = 0;
-  for (const testFunc of tests) {
-    if (startTime < getTimestamp()) {
-      return;
-    }
-    const builtTest = buildTest({
-      collectionID,
-      issuedAt: startTime,
-      testFunc,
-      testID,
-      timeoutInterval,
-    });
-
-    await builtTest();
-    testID += 1;
-  }
-};
-
-const runTestCollections: RunTestCollections = async ({ testCollections }) => {
-  const startTime = updateTimestamp();
   startTestRun(startTime);
 
   let collectionID = 0;
-  for (const collection of testCollections) {
+  for (const collection of testCollection) {
     if (startTime < getTimestamp()) {
       return;
     }
 
     const { tests, runTestsAsynchronously, timeoutInterval } = collection;
-    const runParams: RunTestsParams = {
+    const runParams = {
       collectionID,
       tests,
       startTime,
@@ -170,9 +60,9 @@ const runTestCollections: RunTestCollections = async ({ testCollections }) => {
     });
 
     if (runTestsAsynchronously) {
-      await iterateThroughTests(runParams);
+      await runTestsInOrder(runParams);
     } else {
-      await asynchonouslyRunTests(runParams);
+      await runTestsAllAtOnce(runParams);
     }
 
     if (startTime < getTimestamp()) {
@@ -184,6 +74,7 @@ const runTestCollections: RunTestCollections = async ({ testCollections }) => {
       collectionID,
       endTime,
     });
+
     collectionID += 1;
   }
 
@@ -191,8 +82,15 @@ const runTestCollections: RunTestCollections = async ({ testCollections }) => {
 };
 
 // iterate through tests synchronously
-const runTests: RunTestCollections = async (params) => {
-  runTestCollections(params);
+const runTests: StartLtrTestRun = async (params) => {
+  // create results store
+  const startTime = updateTimestamp();
+
+  await startLtrTestCollectionRun({ ...params, ...{ startTime } });
+  if (startTime < getTimestamp()) {
+    // get state
+    return;
+  }
 };
 
-export { runTests, cancelRun };
+export { Assertions, Test, TestParams, TestCollection, runTests, cancelRun };
